Update retention days and sleep configuration in cleaner; change SleepSeconds type to float64 for better precision.

+++ internal/models/models.go
@@ -109,5 +109,5 @@ type Config struct {
 	BatchSize     int
 	DryRun        bool
 	Verbose       bool
-	SleepSeconds  int
+	SleepSeconds  float64
 }
